==> Step 1 - Obtain Data.ipynb <==
## Import Libraries

## json required . to import user credentials
import json

## Tweepy required to interact with Twitter API
import tweepy


## Pandas required to manipulate data into user-friendly data structure
import pandas as pd

## Set Pandas Display Options

## Set width of pandas dataframe to ensure entire Tweet is displayed
pd.set_option('display.max_colwidth', 3000)

## Retrieve API User Credentials

## Create function to open secret file containing API user credentials
def get_keys(path):
    with open(path) as f:
        return json.load(f)

## Store user credentials in named variables
consumer_key = get_keys("/Users/isobeldaley/.secret/twitter_api.json")['consumer_key']
consumer_secret = get_keys("/Users/isobeldaley/.secret/twitter_api.json")['consumer_secret']
access_token = get_keys("/Users/isobeldaley/.secret/twitter_api.json")['access_token']
access_secret = get_keys("/Users/isobeldaley/.secret/twitter_api.json")['access_secret']

## Provide Twitter API with credentials
auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_secret)

## Set wait_on_rate_limit to True to force queries to wait until rate limit has been replenished
api = tweepy.API(auth, wait_on_rate_limit=True)

## Use Tweepy to Retrieve Tweets to Major UK Mobile Networks

## List of mobile networks for which tweets will be retrieved
networks = ['@VodafoneUK','@EE','@ThreeUK','@O2']

## Create empty list to contain tweets retrieved
list_of_tweets = []


## For loop to iterate through each mobile network for which tweets are to be retrieved
for network in networks:
    
    ## Create a cursor object that will automatically paginate through each search result 
    result = tweepy.Cursor(api.search, q=network, tweet_mode='extended')
    
    ## Retrieve the most recent tweets for each mobile network and add to list_of_tweets
    for tweet in result.items(10000):
        
        dict_ = {'Mobile Network':network,
                'Tweet Created At': tweet.created_at,
                'Tweet Text': tweet.full_text,
                }
        list_of_tweets.append(dict_) 
        

## Manipulate tweets into a DataFrame
df = pd.DataFrame(list_of_tweets)

## Preview dataframe to ensure correct information/formatting etc.
df.head()

## Export Raw Data to Excel for Labelling

## Export raw data to Excel for labelling using Pandas to_excel() function
df.to_excel('unlabelled_data.xlsx')
