==> Step 2 - Scrub Data.ipynb <==
## Import Libraries

## Pandas required to manipulate data into user-friendly data structure
import pandas as pd

## nltk is a leading Python library for Natural Language Processing (NLP) tasks
import nltk

## TextBlob is used for sentiment analysis
from textblob import TextBlob

## stopwords is an nltk function that provides a list of commonly used, low value words (e.g. 'the', 'a' etc.)
from nltk.corpus import stopwords

## Identifies specified patterns within a string 
import re

## Provides list of punctuation
import string

## word_tokenize is an nltk function that splits a string into a list of words
from nltk import word_tokenize

## WordNetLemmatizer is a function that converts a word to its base form (e.g. feet to foot)
from nltk.stem.wordnet import WordNetLemmatizer

## Pickle allows Python objects to be saved for later use, and retrieved
import pickle

## Set Pandas Display Options

## Set width of pandas dataframe to ensure entire Tweet is displayed
pd.set_option('display.max_colwidth', 3000)

## Import Data

## Import data from Excel file (created in Step 1)
df = pd.read_excel('labelled_data.xlsx', index_col=0, usecols='A:E')
df.columns = ['network','datetime','original_tweet','subject']

## Display first five rows of the dataframe
df.head()

## Missing Values

Identify any missing values within the dataframe.  This will prevent further issues with data processing further down the line.

## Use isna().sum() to identify any missing values within the dataframe
df.isna().sum()

As can be seen above, there is only one missing value.  Since it is not possible to predict the contents of a blank tweet, this missing record will be dropped.

## Drop missing values
df.dropna(inplace=True)

## Reset index to ensure sequential index
df = df.reset_index()
df.drop(['index'], axis=1, inplace=True)

## Drop All Retweets from Dataset

To avoid duplicated content, all retweets will be dropped from the dataframe.  Retweets can be identified as those starting with the string "RT".  

## Iterate through each tweet.  For any tweet starting with 'RT', drop it from the dataframe
for i in range(0, len(df)):
    
    if df['original_tweet'][i].startswith('RT'):
        
        df.drop(index=i, inplace=True)

Once all retweets have been removed, the remaining number of observations is as follows:

## Remaining number of tweets
len(df)

## Reset index to ensure sequential index
df = df.reset_index()
df.drop(['index'], axis=1, inplace=True)

## Make Subject Categories Lower Case

The subject will act as a target variable.  To ensure consistent labelling, the different values that this field assumes are checked.

## value_counts() identifies the different values taken on by this field
df['subject'].value_counts(normalize=True)

As can be seen, 'Other' and 'other' are considered different subjects.  So is 'COntract' and 'Contract'.  To prevent this causing issues later on, all values in this field will be made lower case.

## Make all values within the subject field lower case
df['subject'] = df['subject'].str.lower()

To confirm that this has worked, re-run the value_counts() function.

## value_counts() identifies the different values taken on by this field
df['subject'].value_counts(normalize=True)

## Use TextBlob to calculate Sentiment Polarity

One of the objectives of this project is to provide an 'on-the-pulse' measure of customer satisfactions.  In order to support this aim, TextBlob can be used to calculate a sentiment score for each tweet.  The sentiment score ranges from -1 (very negative) to 0 (neutral) to 1 (very positive).  

## Create a new field that indicates sentiment of a tweet using TextBlob
df['sentiment'] = df['original_tweet'].map(lambda text: TextBlob(text).sentiment.polarity)

To check that these sentiment scores are reasonable, 5 randomly selected positive, neutral and negative tweets are checked.

## Print 5 random reviews with the highest positive sentiment (1)
print('5 random reviews with the highest positive sentiment polarity: \n')
positive = df.loc[df.sentiment == 1, ['original_tweet']].sample(5).values
for tweet in positive:
    print(tweet[0])

## Print 5 random reviews with neutral sentiment (0)
print('5 random reviews with neutral sentiment polarity: \n')
neutral = df.loc[df.sentiment == 0, ['original_tweet']].sample(5).values
for tweet in neutral:
    print(tweet[0])

## Print 5 random reviews with the most negative sentiment (-1)
print('5 random reviews with the most negative sentiment polarity: \n')
negative = df.loc[df.sentiment == -1, ['original_tweet']].sample(5).values
for tweet in negative:
    print(tweet[0])

As can be seen from the above, TextBlob has provided a reasonable (but not perfect) score of sentiment.  

## Create List of Stopwords

Commonly used words, such as 'a', 'you', 'the' etc. add little value to an NLP task.  They do not help a model to distinguish between different tweets.  For this reason, it is important that they are removed.  To do this, all of these useless words (known as stopwords) must be identified.  

As a starting point, NLTK provide a comprehensive list of stopwords.  This can then be augmented with punctuation, which also provides no value.  The names of the major phone networks will also be removed.  

## Call the NLTK list of english stopwords
stopwords_list = stopwords.words('english')

## Add common punctuation to this list
stopwords_list += string.punctuation
stopwords_list += ["/n","''", '""', '...', '``',"'",'’','amp']
stopwords_list += ['vodafone', 'three','ee','o2']

Finally, the stopwords list can be previewed.

## Preview the first 20 items on the stopwords list
stopwords_list[0:20]

## Processing Tweeets: Remove Stopwords, Make Lower Case and Tokenize

To pre-process each tweet, the following steps must be completed:

1. Stopwords and any other low value text patterns must be removed
2. All tweets must be made lower case (to ensure any model understands that DOG is the same word as Dog and dog)
3. All tweets must be tokenized.  That is each string must be split into a list.

To make this process most efficient, a function will be created to complete each of these tasks for a given tweet.

## Function to process tweet
def process_tweet(tweet):
    
    ## Remove "@username" from each Tweet
    pattern = '(\w*@\w*)'
    p = re.compile(pattern)
    tweet = p.sub('',tweet)
    
    ## Remove links from each Tweet
    pattern2 = '((http|https)\:\/\/)?[a-zA-Z0-9\.\/\?\:@\-_=#]+\.([a-zA-Z]){2,6}([a-zA-Z0-9\.\&\/\?\:@\-_=#])*'
    p = re.compile(pattern2)
    tweet = p.sub('',tweet)
    
    ## Remove non-english characters
    pattern3 = '([^\x00-\x7A])+'
    p = re.compile(pattern3)
    tweet = p.sub('',tweet)

    ## Tokenize tweet
    tokens = nltk.word_tokenize(tweet)
    
    ## Retain only words that are not in the Stopwords list
    stopwords_removed = [token.lower() for token in tokens if token.lower() not in stopwords_list]
    return stopwords_removed

Having defined this function, each tweet can be processed. The processed tweets are added to an empty list.

## Empty list to contain processed tweets
tokenized_tweets = []

## For loop that iterates through each tweet and processes it, before adding to the list tokenized_tweets
for tweet in list(df['original_tweet']):
    
    tokenized_tweets.append(process_tweet(tweet))
        

To understand the effect of processing a tweet, look at a given tweet before and after processing.

## Tweet before processing
df.loc[15]['original_tweet']

## Tweet after processing
tokenized_tweets[15]

As can be gauged from the above, the processed tweet retains its core meaning, without retaining any surplus-to-requirement words.

## Lemmatization

The next stage of preocessing, it to lemmatize tweets.  Lemmatization is the process of grouping alternative variations of a word, so that they can be interpreted as having a single meaning.  For example, goats becomes goat, and feet becomes foot.  

NLTK provides a comprehensive package to assist with lemmatization called WordNetLemmatizer().  This is used below.

## Define the WordNetLemmatizer() function
lemmatizer = WordNetLemmatizer()

## Create an empty list to contain lemmatized words
lemmatized_tweets = []

## Iterate through each tokenized word to convert to its lemmatized form
for tweet in tokenized_tweets:
    
    lemmatized = []
    
    for word in tweet:
        
        lemmatized.append(lemmatizer.lemmatize(word))
    
    lemmatized_tweets.append(lemmatized)
        
        

## Save Tokenized/Lemmatized Tweets to Modified Dataframe

The tokenized/lemmatized tweets can now be added to the existing dataframe.  They will be added both as tokens and as a string to ensure maximum flexibility during exploratory data analysis and modelling.

## Create a new column in the existing dataframe to contain lemmatized tweets as tokens
df['lemmatized_tweets_tokens'] = lemmatized_tweets

## Create a new column in the existing dataframe to contain the lemmatized tweets as a string
df['lemmatized_tweets_string'] = str()

## For loop to make convert lemmatized/tokenized tweets back into a string
for i in range(0, len(df)):
    
    df['lemmatized_tweets_string'][i] = " ".join(df.loc[i]['lemmatized_tweets_tokens'])

## Preview the first five rows
df.head()

## Split DataFrame into Labelled and Unlabelled Observations

Finally, the dataframes will be split into observations with a subject label, and observations without a subject label.  These will be used as follows in a semi-supervised learning approach:

- The labelled dataframe will be used to build the initial set of models
- A set of labels will then be generated for the unlabelled dataset using this model
- The entire dataset will then be used to rebuild the models, and identify performance improvement.

## Labelled dataframe created and saved using Pickle
df_labelled = df.loc[df['subject']!= " "]
df_labelled.to_pickle('cleaned_labelled_tweets')

## Unabelled dataframe created and saved using Pickle
df_unlabelled = df.loc[df['subject']== " "]
df_unlabelled.to_pickle('cleaned_unlabelled_tweets')
