==> Step 4 - Model Data.ipynb <==
## Import Libraries

All libraries required to create a model capable of classifying tweets by category are imported.  These are described in the comments below.

## Pandas required to manipulate data into user-friendly data structure
import pandas as pd

## Pickle allows Python objects to be saved for later use, and retrieved
import pickle
from sklearn.externals import joblib

## Numpy is used to execute various mathematical functions
import numpy as np

## Matplotlib and Seaborn are both plotting tools used to support datavisualisation
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

## Function to enable random split of data into training and test set
from sklearn.model_selection import train_test_split

## Import the TfidfVectorizer to convert a collection of raw documents to a matrix of TF-IDF features.
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer

## Gridsearch enables the optimal combination of parameters to be selected for a given classifier
from sklearn.model_selection import GridSearchCV

## A number of different classification models
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
import xgboost as xgb

## Metrics to help evaluate the performance of each model
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score, roc_curve, auc

## Set Random Seed

To ensure results are reproducible, a random see is set using Numpy.

## Set a random seed to ensure results are reproducible
np.random.seed(10)

## Set Pandas Display Options

Pandas display settings are chosen to ensure that the full contents of each column can be seen.

## Set width of pandas dataframe to ensure entire Tweet is displayed
pd.set_option('display.max_colwidth', 3000)

## Import Data

The cleaned and labelled tweets are imported from the pre-prepared DataFrame ('cleaned_labelled_tweets').  For more information on how this was created, please refer to **[Step 1 - Obtain Data](https://github.com/isobeldaley/categorising-tweets/blob/master/Step%201%20-%20Obtain%20Data.ipynb)** and **[Step 2 - Scrub Data](https://github.com/isobeldaley/categorising-tweets/blob/master/Step%202%20-%20Scrub%20Data.ipynb)**.

## Import saved dataframe using pickle
df = pd.read_pickle('cleaned_labelled_tweets')

Next, the first five rows of the dataframe are previewed.  

## Preview first five rows of DataFrame
df.head()

As can be seen above, the tweets are categorised by network, subject and sentiment.  They are also stored in three different forms:

- The raw/original tweet
- The cleaned/lemmatized tweet as tokens
- The cleaned/lemmatized tweet as a string

This is to allow greatest flexibility when modelling.

## Split Data into Training & Test Set

To ensure that the models created work well and do not sufer from overfitting, the dataset is split into a training and test set. The training set will be used to build the model, whilst the test set will be used to validate that it works well and can be generalised to new data.

To do this, it is first necessary to define the independent (X) and dependent (y) variables.

## Define the X and y variables
X = df['lemmatized_tweets_string']
y = df['subject']

The **train_test_split()** function can then be used to create this split.  Note that a random_state is specified. This is to ensure results can be reproduced by others.

## Split dataset into training & test set
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2, random_state=213)

## Vectorize Data

Most machine learning algorithms are not able to process raw text directly.  As such, it is first necessary to convert the raw text into vectors of numbers.  

There are a number of different ways of doing this.  This includes using a:

- **Count Vectorizer**, which simply counts the number of times a word appears in a tweet and uses this as its weight.

- **TF-IDF Vectorizer**, which evaulates how important a specified word is in a tweet.  This works by increasing the importance of a word in proportion to the number of times it appears in a particular tweet, but reducing the importance of that word by the frequency the word appears in the entire dataset of tweets.  In this way, it helps the algorithm determine which words are key to categorising a given tweet.  

### Count Vectorizer

Sklearn provides an inbuilt CountVectorizer().  For greatest efficiency, this is used:

## Specify the CountVectorizer, as provided by sklearn
count_vectorizer = CountVectorizer()

Having specified the vectorizer, it is fitted using the training data.  The training data is then transformed.  It is important not to fit using the test data, as this may lead to data leakage from the training to the test data.

## Fit and trasnform the training data using the tf-idf vectorizer
count_X_train = count_vectorizer.fit_transform(X_train)

Finally the test data is also transformed.

## Transform the test data using this vectorizer
count_X_test = count_vectorizer.transform(X_test)

print(count_X_test[1])

### TF-IDF Vectorizer

Sklearn provides an inbuilt TfidfVectorizer().  For greatest efficiency, this is used:

## Specify the tfidfvectorizer, as provided by sklearn
tfidf_vectorizer = TfidfVectorizer()

Having specified the vectorizer, it is fitted using the training data.  The training data is then transformed.  It is important not to fit using the test data, as this may lead to data leakage from the training to the test data.

## Fit and trasnform the training data using the tf-idf vectorizer
tf_idf_X_train = tfidf_vectorizer.fit_transform(X_train)

Finally, the test data is also transformed.

## Transform the test data using this vectorizer
tf_idf_X_test = tfidf_vectorizer.transform(X_test)

## Model Data

Before building and comparing alternative classification models, the following function has been defined. This function identifies the most effective combination of parameters and the best data transformation to enhance model performance for a given classifier.

## Function to identify the optimat dataset and parameters for a given classifier and parameter grid
def best_model_parameters_dataset(classifier, param_grid, datasets):
    
    ## Create a list to contain the dataset, optimal parameters, and score for training and test set
    score_parameters = []
    
    ## Create a for loop which iterates through each dataset and identifies the optimal parameters for the given classifier
    for data in datasets:
        
        gs = GridSearchCV(classifier, param_grid, scoring='accuracy', cv=3)
        gs.fit(data['X_train'], data['y_train'])
        y_test_preds = gs.predict(data['X_test'])
        test_score = accuracy_score(y_test_preds, data['y_test'])
        score_parameters.append({'Dataset':data['name'], 'Training Score':round(gs.best_score_,2), 'Test Score': round(test_score,2), 'Parameters':gs.best_params_})
     
    ## Generate a dataframe that contains the optimal parameters for each dataset
    df = pd.DataFrame(score_parameters)
    df.sort_values(by=['Test Score', 'Training Score'], inplace=True, ascending=False)
    
    return df

A number of different classification models will be assessed for their suitability for the task of predicting the category a given tweet relates to.  These are:

- K Nearest Neighbors
- Naive Multinomial Bayes Classifier
- Multinomial Logistic Regression
- Random Forest Classifier
- XG Boost
- Support Vector Machine (SVM)

Each model will be created using the two datasets specified above:

- Dataset transformed by tf-idf vectorization
- Dataset transformed by count vectorization

To make this task more efficient, a list containing all training and test datasets is created.

## Create a list of datasets.  Each item is a dictionary detailing training/test datasets
datasets = [{'name': 'tf_idf','X_train': tf_idf_X_train, 'y_train':y_train, 'X_test': tf_idf_X_test, 'y_test':y_test},
           {'name':'count','X_train':count_X_train,'y_train':y_train, 'X_test': count_X_test, 'y_test':y_test}]

### K Nearest Neighbours

The K Nearest Neighbours (KNN) works by identifying a specified number of similar observations ('nearest neighbours) based on a specified distance metric, and then providing a classification based on the majority classification of the identified 'neighbours'.  

## Specify the classifier, in this case K nearest neighbours
knn = KNeighborsClassifier()

## Define the parameter grid
knn_param_grid = {'n_neighbors':[5,20,40,50,60],
              'metric': ['manhattan', 'euclidean','minkowski'],
              'weights': ['uniform', 'distance']
             }

## Run the best_model_parameters_dataset() function to identify the optimal dataset and parameters to use
best_model_parameters_dataset(knn, knn_param_grid, datasets)

From teh above, it appears that the best performance is achieved when the dataset that has been transformed by a CountVectorizer() is used, and the following parameters are specified:

- **Distance Metric**: Euclidean
- **Number of Neighbours**: 5
- **Weights**: Distance

To assess the performance of this model in morde detail, the model will be created using these optimal parameters.

## Create a KNN classifier using the optimal parameters specified above
knn = KNeighborsClassifier(metric='euclidean', n_neighbors=5, weights='distance')

## Fit the model using the count vectorized dataset
knn.fit(count_X_train, y_train)


Next, predictions can be generated using the model for both the test and training set.  These will be used to calculate a number of metrics to assess performance of the model.

knn_preds_train = knn.predict(count_X_train)
knn_preds_test = knn.predict(count_X_test)

#### Classification Report

The classification report provides the following key metrics:
- **Precision**: Defined as the number of times the model correctly assigned a given classification,  as a proportion of all observations with that predicted classification.  A low precison suggests a high rate of false positives.  
- **Recall**: Defined as the number of times the model correctly assigned a given classification,  as a proportion of all observations with that actually had that classification.  A low recall suggests a high rate of false negatives.  
- **f1-score**: The harmonic mean of precision and recall.  
- **Accuracy**: The proportion of all observations that were correctly classified

Note that there is an inverse relationship between precision and recall.  Balance is therefore important.  

## Print the classification report for the training dataset
print(classification_report(y_train, knn_preds_train))

## Print the classification report for the test daaset
print(classification_report(y_test, knn_preds_test))

The following observations can be made:

- There is a significant mismatch between the weighted average of accuracy of the training and test set (99% v. 52%).  This is indicative of overfitting, and suggests the model does not generalize well to new data.  

- Whilst there is good balance between precision and recall for the training data, this is not the case for the test data.  With the exception of the 'other' category, precision is notably larger than recall.  This suggests that there is an issue with false negatives.

The performance of this model is too low to take it forward to production.  Alternatives must be sought.

### Multinomial Naive Bayes Classifier

Next, the Multinomial Naive Bayes Classifier is assessed for suitability.  This classifier is based on Bayes theorem.

## Specify the classifier, in this case naive bayes
nb = MultinomialNB()

## Create a parameter grid to identify optimal parameters
nb_param_grid = {'alpha':[0.5,0.8,1]}

## Run the best_model_parameters_dataset() function to identify the optimal dataset and parameters to use
best_model_parameters_dataset(nb, nb_param_grid, datasets)

To assess the performance further, this model will be created using the optimal combination of parameters:
- Count Vectorized dataset
- Alpha = 0.8

## Create a multinomial NB classifier using the optimal parameters specified above
mnb = MultinomialNB(alpha=0.8)

## Fit the model using the count vectorized dataset
mnb.fit(count_X_train, y_train)


Next, predictions can be generated using the model for both the test and training set. These will be used to calculate a number of metrics to assess performance of the model.

## Generate predictions for the MNB model
mnb_preds_train = mnb.predict(count_X_train)
mnb_preds_test = mnb.predict(count_X_test)

#### Classification Report

## Print the classification report for the training dataset
print(classification_report(y_train, mnb_preds_train))

## Print the classification report for the test dataset
print(classification_report(y_test, mnb_preds_test))

From the above it can be seen that:

- There is still evidence of overfitting, as accuracy predictions using the training data are greater than those using the test data.  However, this shows better balance than under the K Nearest Neighbours classifier.
- There is better balance between precision and recall for the training and test data.  

### Multinomial Logistic Regression

Logistic Regression is a classification algorithm that employs Maximum Likelihood Estimation to generate a model capable of dividing observations into different groups.

## Specify the classifier, in this case LogisticRegression()
logreg = LogisticRegression(random_state=55, max_iter=15000, multi_class='multinomial')

## Create a parameter grid to identify optimal parameters
logreg_param_grid = {'C':[1,2,10],
                     'class_weight': ['balanced', None],
                     'solver':['newton-cg', 'sag', 'saga','lbfgs']}

best_model_parameters_dataset(logreg, logreg_param_grid, datasets)

The above indicates that the following combination of parameters lead to optimal performance:

- **Dataset**: the data vectorised using tf-idf yields marginally better performance
- **C**: 2
- **class_weight**: balanced
- **solver**: newton-cg

The model is therefore fitted using these parameters.

## Fit the model using optimal dataset and parameters
logreg_ = LogisticRegression(random_state=213, solver='newton-cg', max_iter=15000, C=2, class_weight='balanced')
logreg_.fit(tf_idf_X_train, y_train)

To assess the performance of each model, predictions are generated for teh training and test dataset.

## Create predictions for the training and test dataset
logreg_preds_train = logreg_.predict(tf_idf_X_train)
logreg_preds_test = logreg_.predict(tf_idf_X_test)

#### Classification Report

Again, the classification report is generated for the training and test data.

## Print the classification report for the training dataset
print(classification_report(y_train, logreg_preds_train))

## Print the classification report for the test daaset
print(classification_report(y_test, logreg_preds_test))

There are similar issues with overfitting with this model.  The accuracy of predictions using the training data is about 17% higher than the model built using the test data.  

**Note**: By default, an l2 penalty is included within a logistic regression model solved using 'newton-cg'.  It is unfortunately not possible to make any enhancements to this penalty in order to improve the issue of overfitting.  

To see if this can be improved, an l2 penalty will be introduced.  This provides one mechanism to overcome overfitting.

It can also be seen from the above that the model performs poorly when classifying tweets as 'broadband' and 'contract', but performs much better when classifying into the other categories.

### Random Forest

A random forest is considered as the next possible classifier.  Random forests work by creating a number of different decision trees (specified by n_estimators) and then outputting the mode of the predictions made by each decision tree.

## Define the classifier to be used, in this case RandomForestClassifier(), specify a random_state
## so that the results are reproducible
forest = RandomForestClassifier(random_state=55)

## Specify the parameter grid to be assessed
forest_param_grid = {'n_estimators': [75,150,300,450],
                    'criterion': ['gini', 'entropy'],
                  'max_depth':[None, 5, 10, 15],
                  'class_weight': ['balanced', None],
                  'bootstrap': [True, False]
             }

## Run the best_model_parameters_dataset() function to identify the optimal dataset and parameters to use
best_model_parameters_dataset(forest, forest_param_grid, datasets)

As can be seen from the above, the following combinations of parameters yield optimal performance:
- **Dataset**: Data vectorized using tf-idf vectorization
- **bootstrap**: False
- **class_weight**: balanced
- **criterion**: gini
- **max_depth**: None
- **n_estimators**:450

To investigate the performance of this model further, it is created below.

## Define the random forest classifier
forest = RandomForestClassifier(bootstrap=False, criterion='gini', max_depth=None, 
                                n_estimators=450, random_state=213, class_weight='balanced')

## Fit the model to the tf_idf data
forest.fit(tf_idf_X_train, y_train)

## Generate predictions using this model
forest_preds_train = forest.predict(tf_idf_X_train)
forest_preds_test = forest.predict(tf_idf_X_test)

#### Classification Report

The classification report is generated.

## Generate the classification report for the training data
print(classification_report(y_train, forest_preds_train))

## Generate the classification report for the testing data
print(classification_report(y_test, forest_preds_test))

In the above, there is evidence of marked overfitting (accuracy is 99% with training data, but 75% with test data).  The balance between precision and recall is mixed.

### Support Vector Machine

Next, the Support Vector Machine (SVM).  This model attempts to find the decision boundary which maximises the distance between the boundary and the training observatiopns.  This model includes a parameter (C), which specifies the balance between finding this optimal boundary for most datapoints, and misclassifying observations.

## Specify the classification model, in this case a support vector machine
svm = SVC(gamma='auto', random_state=55)

## Specify the parameter grid to be used during the GridSearchCV
svm_param_grid = {'C':[1,5,10],
                'class_weight':['balanced', None]}

## Run the best_model_parameters_dataset() function to identify the optimal dataset and parameters to use
best_model_parameters_dataset(svm, svm_param_grid, datasets)

As can be seen from the above, performance of the SVM is weak relative to some of the other models (e.g. Multinomial Logistic Regression).  For this reason, we will not progress with this classifier.

### XG Boost

The XG Boost model is not part of the sklearn library. Therefore GridSearch will not be performed for this model. However, the performance of this model will be compared both using the tf-idf dataset and the count vectorized dataset.

#### XG Boost with TF-IDF Dataset

## Specify the classifier, in this case XG Boost
boost = xgb.XGBClassifier()

## Fit the model using the training data
boost.fit(tf_idf_X_train, y_train)

## Create predictions for the training and test datasets
boost_preds_train = boost.predict(tf_idf_X_train)
boost_preds_test = boost.predict(tf_idf_X_test)

## Print the accuracy score for XG Boost using training data
accuracy_score(y_train, boost_preds_train)

## Print the accuracy score for XG Boost using test data
accuracy_score(y_test, boost_preds_test)

#### XG Boost with Count Vectorized Dataset

## Specify the classifier, in this case XG Boost
boost = xgb.XGBClassifier()

## Fit the model using the training data
boost.fit(count_X_train, y_train)

## Create predictions for the training and test datasets
boost_preds_train = boost.predict(count_X_train)
boost_preds_test = boost.predict(count_X_test)

## Print the accuracy score for XG Boost using training data
accuracy_score(y_train, boost_preds_train)

## Print the accuracy score for XG Boost using test data
accuracy_score(y_test, boost_preds_test)

#### Assessment

As can be seen from the above, the XG Boost performs slightly better when using the tf-idf dataset.  

## Label Unlabelled Data

Having built the initial model, it is possible to investigate whether predicting labels for the unlabelled data, and then  re-running this model to include the predicted categories for the unlabelled data, will improve performance.

Before doing this, it is first necessary to import the unlabelled data, which has been stored in a Pandas DataFrame using pickle.

## Import the unlabelled data using Pickle
unlabelled_df = pd.read_pickle('cleaned_unlabelled_tweets')

## Display the first five rows of the DataFrame
unlabelled_df.head()

Next, it is necessary to vectorize the unlabelled tweets using the tf_idf vectorizer.

## Transform unlabelled tweets using the tfidf_vectorizer
tf_idf_unlabelled_X_train = tfidf_vectorizer.transform(unlabelled_df['lemmatized_tweets_string'])

Once this is complete, it is possible to predict categorise for the unlabelled data using the chosen model.  The logistic regression model is chosen as it offers a good balance between accuracy for the training and test data.

## Make predictions for the unlabelled tweets
unlabelled_y_train = logreg.predict(tf_idf_unlabelled_X_train)

Finally, the unlabelled tweets and the predicted categories can be combined into a single dataset.

## Create a combined list of tweets (X)
all_X_train = list(X_train)
all_X_train.extend(unlabelled_df['lemmatized_tweets_string'])

## Vectorize the complete list of tweets
tf_idf_all_X_train = tfidf_vectorizer.transform(all_X_train)

## Create a combined list of categories (y)
all_y_train = np.append(values=np.array(y_train), arr=unlabelled_y_train)

## Re-Run Model

Finally, a new model can be created using the enlarged dataset.  To ensure assessment is comparable to prior models, the same test dataset will be used as before.  A multinomial logistic regression model will be used.

## Create a list containing the combined dataset
complete_datasets = [{'name': 'complete_tf_idf','X_train': tf_idf_all_X_train, 'y_train':all_y_train, 'X_test': tf_idf_X_test, 'y_test':y_test}]


## Specify the classifier, in this case LogisticRegression()
logreg = LogisticRegression(random_state=55, max_iter=15000, multi_class='multinomial')

## Create a parameter grid to identify optimal parameters
logreg_param_grid = {'C':[1,2,10],
                     'class_weight': ['balanced', None],
                     'solver':['newton-cg', 'sag', 'saga','lbfgs']}

## Run the best_model_parameters_dataset() function to identify the optimal dataset and parameters to use
best_model_parameters_dataset(logreg, logreg_param_grid, complete_datasets)

The results above suggest that the model deteriorates when using unlabelled data categorised by the original model.  This will therefore not be taken further.  

## Save Model using Pickle

The chosen logistic regression model (run using only the labelled data) is saved using a Pickle object.  This model will be used to generate predictions in the next step **Step 5 - Interpret Results**.

## Use Pickle and joblib to save the Pickle object
joblib.dump(logreg_, 'model.pkl')

pickle.dump(X_test, open( "X_test.pkl", "wb" ))
pickle.dump(y_test, open( "y_test.pkl", "wb" ))
pickle.dump(tf_idf_X_test, open("tf_idf_X_test.pkl", "wb"))

print(tf_idf_X_test[0])
