==> Step 5 - Interpret Model.ipynb <==
## Import Libraries

All libraries required to create a model capable of classifying tweets by category are imported.  These are described in the comments below.

## Pickle allows Python objects to be saved for later use, and retrieved
import pickle

## Works with Pickle to load sklearn models
from sklearn.externals import joblib

## Pandas required to manipulate data into user-friendly data structure
import pandas as pd

import numpy as np

## Set Pandas Display Options

Pandas display settings are chosen to ensure that the full contents of each column can be seen.

## Set width of pandas dataframe to ensure entire Tweet is displayed
pd.set_option('display.max_colwidth', 3000)

## Create Colour Class to Format Text

A class is created to enable text to be printed in bold and with an underline.  This is used later in this workbook.

## Define class to enable text to be emboldened and underlined
class colour:
   BOLD = '\033[1m'
   UNDERLINE = '\033[4m'
   END = '\033[0m'

## Load Saved Model

The logistic regression model, compiled during **[Step 4 - Model Data](https://github.com/isobeldaley/categorising-tweets/blob/master/Step%204%20-%20Model%20Data.ipynb)** is loaded using Pickle.

## Load logistic regression model using Pickle
model = joblib.load('model.pkl')

## Load Test Data

Next, to understand how this model performs when classifying tweets, the test data (obtained during **[Step 4 - Model Data](https://github.com/isobeldaley/categorising-tweets/blob/master/Step%204%20-%20Model%20Data.ipynb)** is loaded using Pickle.  

## Load X_test, y_test and X_test vectorized using tf-idf vectorization in step 4
X_test = pickle.load(open("X_test.pkl", "rb"))
y_test = pickle.load(open("y_test.pkl", "rb"))
tf_idf_X_test = pickle.load(open("tf_idf_X_test.pkl", "rb"))

In addition, the table containing the original tweet and sentiment rating is loaded (this was created in **[Step 2 - Scrub Data](https://github.com/isobeldaley/categorising-tweets/blob/master/Step%202%20-%20Scrub%20Data.ipynb)**).

## Load table containing original tweet and sentiment rating
df = pd.read_pickle('cleaned_labelled_tweets')

## Select a Random Sample of Tweets

To understand how well this model works, a sample of 10 tweets are selected from the test data.  Predicted categories will then be generated for these tweets.

## Select a sample of 10 tweets from the test dataset (NB: What is shown below if the cleaned/lemmatized tweet)
X_test[215:225]

To provide a complete comparison, the original tweet, alongside the TextBlob sentiment rating is extracted:

## Select original tweet and sentiment for each of the tweets above
df = df.loc[[9460,10443,6881,5787,778,9793,5627,929,5779,5614],['original_tweet','sentiment']]

## Process Tweets Through Model

Next, the model is used to generate predictions for each of the selected test tweets.

## Isolate the vectorized test tweets for which predictions will be made
tweets_to_process = tf_idf_X_test[215:225]

## Generate predictions for each of the test tweets using the model
predictions = model.predict(tweets_to_process)

## Compare Predicted Category & Sentiment to Original Tweet

Before comparing the original tweet with the predicted category and sentiment, the results are combined in a single dataframe.

## Combine predicted category/sentiment and original tweet in single dataframe
results_df = pd.DataFrame({'Tweet': df['original_tweet'], 'Actual Category': y_test[215:225], 
                   'Predicted Category': predictions, 
                   'Predicted Sentiment':df['sentiment']})

Then, the results are printed for comparison.

## Display results 
results_df

The performance of the model appears to be reasonable.  On 8 out of 10 occasions the category is correctly predicted. 

There are two tweets that are labelled "customer service" when the human has labelled them "contract" and "other".  However, upon deeper reading of the tweet, it could be argued that either or both could have been labelled "customer service".

Performance of the sentiment rating is more difficult to judge.  The first tweet is clearly negative, and yet has a mildly positive sentiment rating of 0.18.  Meanwhile, the majority of tweets have been given a neutral rating.  More work may be needed to fine tune this element of the project.

## On-the-Pulse Measure of Customer Satisfaction

At the outset, it was stated that an objective of this project was to provide mobile networks with an on-the-pulse measure of customer satisfaction.  To understand how this would work, measures are provided for the demo data above.

### Subject

First, a dataframe is created showing the proportion of tweets relating to each subject category.

## Create a dataframe to show the proportion of tweets relating to each category
categories = pd.DataFrame({'Proportion of Tweets (%)':results_df['Predicted Category'].value_counts(normalize=True)})
categories['Proportion of Tweets (%)'] = (categories['Proportion of Tweets (%)']*100).astype(int)

## Display the proportion of tweets relating to each category
print(colour.BOLD+colour.UNDERLINE +'Proportion of Tweets by Category'+colour.END)
categories

### Sentiment

Next, the mean sentiment for all tweets is assessed:

## Print the mean sentiment for all tweets
print(colour.BOLD+colour.UNDERLINE +'Mean Sentiment of Tweets'+colour.END)
round(np.mean(results_df['Predicted Sentiment']),2)

## Recommendations

Having demonstrated its potential, we recommend that this model is used in the following ways:

1. To pre-categorize tweets.  This will enable all questions/issues to be directed to the correct person/department and handled efficiently.

2. To provide a measure of the proportion of tweets relating to each subject category.  This will help mobile network operators to quickly identify and address specific (e.g. persistent problems with customer service, or poor network coverage).

3.  To provide an on-the pulse measure of customer satisfaction, by considering the distribution of sentiment ratings.

This approach will enable more effective management of customer communications via social media, rapid assessment of customer sentiment/issues.


## Future Work

There are a number of areas that would merit further investigation in the future.  These are detailed below.


### Labelling & Categorisation

For this project, we had access to 4,377 labelled tweets.  It is highly likely that model performance could be improved if the labelled dataset was significantly expanded.  Given the appropriate time and resources, it is recommended that this is done prior to deploying the model.  

Moreover, it was noted that a large proportion of tweets were categorised as "Other".  This suggests that more detailed categorisation may be needed.  

### Tweets & Replies

This model considered tweets individually.  However, tweets are often part of a longer conversation between a service agent and a customer.  It may be useful to broaden the model to consider original tweets alongside any replies.  This would help operators understand how many tweets and how much time was required to resolve an issue/question.  This is a crucial component of customer service.  

### Sentiment Analysis

Finally, it was noted that the sentiment analysis was often inaccurate.  An alternative means of modelling sentiment should be considered to improve performance of this element of the model.
